ml/scraping/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
redis==5.0.1
httpx==0.25.2
lxml==4.9.3
cython==3.0.6
boto3==1.34.34
python-dotenv==1.0.1
setuptools==78.1.0
langdetect==1.0.9
requests==2.31.0
tqdm==4.66.2